mock-db-client/pyproject.toml
[project]
name = "mock-db-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mysql-connector-python>=9.3.0",
    "psycopg2>=2.9.10",
]
